First iteration of pipes implemented and working! Needs testing. Not normed. Now implementing redirections.
