+++ srcs/pipeline.c
@@ -6,11 +6,11 @@
 /*   By: smagdela <smagdela@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/15 12:47:56 by smagdela          #+#    #+#             */
-/*   Updated: 2022/03/25 13:00:26 by smagdela         ###   ########.fr       */
+/*   Updated: 2022/03/25 15:05:21 by smagdela         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "parsing.h"
+#include "minishell.h"
 
 /*
 Reproduce the behavior of multiple pipes, adding file redirectors too.
@@ -29,8 +29,8 @@ static void	redirection_handler(t_token *token_list, char **envp, t_data *data)
 				perror("pipe failed.");
 				exit_ms(NULL, data);
 			}
-			tmp->previous->out = tmp->pipefd[0];
-			tmp->next->in = tmp->pipefd[1];
+			tmp->previous->out = tmp->pipefd[1];
+			tmp->next->in = tmp->pipefd[0];
 		}
 		tmp = tmp->next;
 	}
@@ -38,10 +38,7 @@ static void	redirection_handler(t_token *token_list, char **envp, t_data *data)
 	while (tmp)
 	{
 		if (tmp->type == WORD)
-		{
-			
 			fork_exec(tmp, envp, data);
-		}
 		tmp = tmp->next;
 	}
 }
@@ -54,7 +51,7 @@ bool	executor(t_token *token_list, char **envp, t_data *data)
 	if (token_list == NULL)
 		return (false);
 	if (token_list->next == NULL)
-		parse_line(token_list->cmd, envp, data);
+		fork_exec(token_list, envp, data);
 	else
 		redirection_handler(token_list, envp, data);
 	if (token_list)
